rename pkg/specs -> pkg/compliance
