+++ pkg/compliance/aws-cis-1.2.yaml
