+++ pkg/compliance/loader.go
@@ -1,4 +1,4 @@
-package specs
+package compliance
 
 import (
 	"embed"
