+++ pkg/compliance/loader_test.go
@@ -1,4 +1,4 @@
-package specs
+package compliance
 
 import (
 	"os"
