+++ pkg/compliance/rke2-cis-1.24.yaml
